Importing Spike files into Matlab struct
Currently this is customized for how Krista was initially doing things manually to start making a good script to import as desired. Could be made more general for this master branch so that each person could customize their own personal from the master. Or each person could keep a personal master in this folder so that others can access different specific usages.